backend/main_app/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
psycopg2-binary==2.9.10
django-cors-headers==4.7.0
python-decouple==3.8
drf-yasg==1.21.10